requirements.txt
fastapi
jinja2
requests
uvicorn
python-dotenv
python-jose[cryptography]
python-multipart